requirements.txt
python-dotenv==1.0.1
SQLAlchemy==1.4.54
pandas==2.0.3
psycopg2-binary==2.9.10
